+++ test/intl/display_names_test.exs
@@ -33,8 +33,37 @@ defmodule Intl.DisplayNamesTest do
                Intl.DisplayNames.of(:year, type: :date_time_field, locale: :en)
     end
 
+    test "date_time_field accepts the JS field spellings" do
+      assert {:ok, "week"} =
+               Intl.DisplayNames.of(:week_of_year, type: :date_time_field, locale: :en)
+
+      assert {:ok, "week"} =
+               Intl.DisplayNames.of("weekOfYear", type: :date_time_field, locale: :en)
+
+      assert {:ok, "time zone"} =
+               Intl.DisplayNames.of(:time_zone_name, type: :date_time_field, locale: :en)
+
+      assert {:ok, "time zone"} =
+               Intl.DisplayNames.of("timeZoneName", type: :date_time_field, locale: :en)
+    end
+
+    test "date_time_field accepts the Localize field names" do
+      assert {:ok, "week"} = Intl.DisplayNames.of(:week, type: :date_time_field, locale: :en)
+      assert {:ok, "time zone"} = Intl.DisplayNames.of(:zone, type: :date_time_field, locale: :en)
+
+      assert {:ok, "day of the week"} =
+               Intl.DisplayNames.of(:weekday, type: :date_time_field, locale: :en)
+
+      assert {:ok, "AM/PM"} =
+               Intl.DisplayNames.of(:day_period, type: :date_time_field, locale: :en)
+    end
+
     test "missing type returns error" do
       assert {:error, _} = Intl.DisplayNames.of("US")
+    end
+
+    test "unknown type returns an error rather than raising" do
+      assert {:error, %ArgumentError{}} = Intl.DisplayNames.of("US", type: :nonsense)
     end
   end
 end